rgw_sal_motr: [CORTX-32482] don't multiply delete markers
On deletion of an object in the versioned bucket, don't create
additional delete markers if there is one created already and
it's the latest version of the object.

This behaviour was changed recently at 09c5da2aa2b, but after
discussion, it seems more appropriate performance-wise to keep
the old behavior.

Note: AWS S3 specifies the creation of additional delete markers,
(see the note at
https://docs.aws.amazon.com/AmazonS3/latest/userguide/ManagingDelMarkers.html)
but we prefer to behave similar to RGW in this particular case.

Signed-off-by: Andriy Tkachuk <andriy.tkachuk@seagate.com>